fix: Application Sign Up and Sign In fields for organisation (#86)
The Application entity found at org.casbin.casdoor.entity.Application is missing required fields that are needed to correctly load the OAuth2 page and admin dashboard.

Fields missing are:

- Sign Up Items
- Sign In Items
- Sign In Methods

Due to these missing fields the database is holding null. This results with a front-end error as the UI is expecting an array (even if it's an empty array) in comparison to null for both the OAuth2 login page, and the Admin Dashboard.